territorios/requirements.txt
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==6.0.1
Django==5.0.1
etelemetry==0.3.1
exceptiongroup==1.2.0
filelock==3.13.3
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.6
isodate==0.6.1
looseversion==1.3.0
lxml==5.2.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathlib==1.0.1
pillow==10.2.0
prov==2.0.0
pydot==2.0.0
PyMuPDF==1.24.0
PyMuPDFb==1.24.0
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-telegram-bot==21.0.1
pytz==2024.1
pyxnat==1.6.2
rdflib==7.0.0
reportlab==4.1.0
requests==2.31.0
scipy==1.12.0
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
traits==6.3.2
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.1
